Reorganized the printed_parts directory and added auxiliary files
I’ve reorganized the printed_parts directory, using subdirectories for
helping users choose which parts to download and print.
Also, I have added the auxiliary files I’ve used in the design of the
printer parts, so anyone can fully modify the objects. These files are
marked with aux_ in the filename.