Big commit to reach version 1.0
This commit uploads a lot of changes in the machine and the repository
structure to reach version 1.0 of the M Prime One project.

We have make some improvements and evolutions on the machine:

* New frame, directly compatible with the PTSone extruder (the new
default extruder).
* New 3D printed extruder, interchangeable with the PTSone all metal
extruder. This extruder however hasn’t been fully tested for
reliability and it is shared as a starting point for the users that
don’t want to use a non printed extruder and want to develop a new one.
* New x_axis_end part that uses magnets with bolts so it is not
necessary to glue the magnets to the printed part anymore.
* Added 5mm on the X axis to assure 200mm travel (affects the bed and
rod files/specs).

And changes in the files and the repository:

* New repository organization, now each part has its own directory.
* New 3D assembly model made from scratch with many improvements: E3D
hotend, timing belts, zip ties… This is the starting point for the
assembly manual that will be released next.
* We have uploaded and properly attributed every part of the machine,
you can find and use all the components of the 3D model.
* Added a fabrication file document to help users find all the .stl and
.dxf files in the new file structure.
* Added .dxf fabrication files for the printing surface.